Tidy trygdetid regler and add comments.
